pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fsmgen"
version = "0.1.0"
description = "Generate Verilog state machines from YAML"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
fsmgen = "fsmgen.cli:cli"

[tool.setuptools.packages.find]
include = ["fsmgen*"]
